Update matrix.lua
Fix to matrix.copy() function. It was using row length to set column loop length and column length to set row loop length, resulting in the copy always being a square matrix. I have switched the two culprit lines to make the out put be an identical copy of the input instead of the largest square matrix contained in the input matrix.